// github/types/paths/user/repos.types.ts
export type reposListForAuthenticatedUser = (
  $: OmitValueWhenNever<{
    query: reposListForAuthenticatedUser_Query;
    querystring: never;
    path: never;
    headers: never;
    cookie: never;
    body: never;
    context: Context;
    response: ResponseBuilderFactory<{
      200: {
        headers: {
          Link: { schema: string };
        };
        requiredHeaders: never;
        content: {
          "application/json": {
            schema: Array<repository>;
          };
        };
        examples: {
          default: unknown;
        };
      };
      304: not_modified;
      401: requires_authentication;
      403: forbidden;
      422: validation_failed;
    }>;
    x: WideOperationArgument;
    proxy: (url: string) => COUNTERFACT_RESPONSE;
    user: never;
    delay: (milliseconds: number, maxMilliseconds?: number) => Promise<void>;
    version: never;
  }>,
) => MaybePromise<COUNTERFACT_RESPONSE>;
// ...
export type reposListForAuthenticatedUser_Query = {
  /**
   * Limit results to repositories with the specified visibility.
   */
  visibility?: "all" | "public" | "private";
  /**
   * Comma-separated list of values. Can include:
   *  * `owner`: Repositories that are owned by the authenticated user.
   *  * `collaborator`: Repositories that the user has been added to as a collaborator.
   *  * `organization_member`: Repositories that the user has access to through being a member of an organization. This includes every repository on every team that the user is on.
   */
  affiliation?: string;
  /**
   * Limit results to repositories of the specified type. Will cause a `422` error if used in the same request as **visibility** or **affiliation**.
   */
  type?: "all" | "owner" | "public" | "private" | "member";
  /**
   * The property to sort the results by.
   */
  sort?: "created" | "updated" | "pushed" | "full_name";
  /**
   * The order to sort by. Default: `asc` when using `full_name`, otherwise `desc`.
   */
  direction?: "asc" | "desc";
  per_page?: number;
  page?: number;
  since?: string;
  before?: string;
};